refactor(providers): extract queryClient + persister from QueryProvider
Splits the Phase 2 provider into:
- queryClient.ts — React-free singleton + persister + shared constants
  (CACHE_MAX_AGE, CACHE_VERSION, CACHE_KEY)
- queryProvider.tsx — thin React wrapper around PersistQueryClientProvider

Prep for Phase 1.B offscreen document: the offscreen runtime has no
React, can't import a .tsx pulling in ReactNode. Offscreen will
consume the same singleton from queryClient.ts, so writes from both
contexts land in the same shared chrome.storage.local entry — the
cross-context cache bus the Phase 1 plan describes.

Zero behavior change for the popup — queryProvider.tsx still exports
the same QueryProvider component + re-exports queryClient for
consumers that grabbed it from the old path.
